+++ src/DTO/OrderData.php
@@ -383,46 +383,6 @@ class OrderData implements JsonSerializable
         $this->payment = $payment;
     }
 
-    /**
-     * @return string
-     */
-    public function getShippingStreetAndNumber()
-    {
-        return $this->shippingStreetAndNumber;
-    }
-
-    /**
-     * @param string $shippingStreetAndNumber
-     *
-     * @return self
-     */
-    public function setShippingStreetAndNumber($shippingStreetAndNumber)
-    {
-        $this->shippingStreetAndNumber = $shippingStreetAndNumber;
-
-        return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getBillingStreetAndNumber()
-    {
-        return $this->billingStreetAndNumber;
-    }
-
-    /**
-     * @param string $billingStreetAndNumber
-     *
-     * @return self
-     */
-    public function setBillingStreetAndNumber($billingStreetAndNumber)
-    {
-        $this->billingStreetAndNumber = $billingStreetAndNumber;
-
-        return $this;
-    }
-
     public function jsonSerialize()
     {
         $lines = [];
@@ -436,24 +396,26 @@ class OrderData implements JsonSerializable
                 'value' => (string) $this->getAmount()->getValue(),
             ],
             'billingAddress' => [
-                'organizationName' => ltrim($this->getBillingAddress()->company),
-                'streetAndNumber' => substr($this->getBillingStreetAndNumber(), 0, 100),
-                'city' => $this->getBillingAddress()->city,
-                'postalCode' => $this->getBillingAddress()->postcode,
-                'country' => (string) Country::getIsoById($this->getBillingAddress()->id_country),
-                'givenName' => $this->getBillingAddress()->firstname,
-                'familyName' => $this->getBillingAddress()->lastname,
-                'email' => $this->getEmail(),
+                'organizationName' => $this->cleanUpInput($this->getBillingAddress()->company),
+                'streetAndNumber' => $this->cleanUpInput($this->getBillingAddress()->address1),
+                'streetAdditional' => $this->cleanUpInput($this->getBillingAddress()->address2),
+                'city' => $this->cleanUpInput($this->getBillingAddress()->city),
+                'postalCode' => $this->cleanUpInput($this->getBillingAddress()->postcode),
+                'country' => $this->cleanUpInput(Country::getIsoById($this->getBillingAddress()->id_country)),
+                'givenName' => $this->cleanUpInput($this->getBillingAddress()->firstname),
+                'familyName' => $this->cleanUpInput($this->getBillingAddress()->lastname),
+                'email' => $this->cleanUpInput($this->getEmail()),
             ],
             'shippingAddress' => [
-                'organizationName' => ltrim($this->getShippingAddress()->company),
-                'streetAndNumber' => substr($this->getShippingStreetAndNumber(), 0, 100),
-                'city' => $this->getShippingAddress()->city,
-                'postalCode' => $this->getShippingAddress()->postcode,
-                'country' => (string) Country::getIsoById($this->getShippingAddress()->id_country),
-                'givenName' => $this->getShippingAddress()->firstname,
-                'familyName' => $this->getShippingAddress()->lastname,
-                'email' => $this->getEmail(),
+                'organizationName' => $this->cleanUpInput($this->getShippingAddress()->company),
+                'streetAndNumber' => $this->cleanUpInput($this->getShippingAddress()->address1),
+                'streetAdditional' => $this->cleanUpInput($this->getShippingAddress()->address2),
+                'city' => $this->cleanUpInput($this->getShippingAddress()->city),
+                'postalCode' => $this->cleanUpInput($this->getShippingAddress()->postcode),
+                'country' => $this->cleanUpInput(Country::getIsoById($this->getShippingAddress()->id_country)),
+                'givenName' => $this->cleanUpInput($this->getShippingAddress()->firstname),
+                'familyName' => $this->cleanUpInput($this->getShippingAddress()->lastname),
+                'email' => $this->cleanUpInput($this->getEmail()),
             ],
             'redirectUrl' => $this->getRedirectUrl(),
             'webhookUrl' => $this->getWebhookUrl(),
@@ -474,5 +436,19 @@ class OrderData implements JsonSerializable
         }
 
         return $result;
+    }
+
+    private function cleanUpInput($input)
+    {
+        $defaultValue = 'N/A';
+
+        if (!isset($input) || empty($input)) {
+            return $defaultValue;
+        }
+
+        $input = ctype_space($input) ? $defaultValue : $input;
+        $input = ltrim($input);
+
+        return substr($input, 0, 100);
     }
 }