+++ changelog.md
@@ -7,6 +7,8 @@
 + Removed reference variables from translations.
 + Fixed issue where not secure credit card would sometimes throw error on checkout.
 + Fixed issue where on php5.6 order api would not change status in some cases.
++ Added street additional information when creating order.
++ Fixed issue where country restriction was using delivery address and not invoice.
 
 ## Changes in release 4.5.2 ##
 + Fixed issue where credit card payment would show id not found instead of order confirmation page when order is paid.